refactor: Optimize Payment Gateway
The authority retrieval is now situated within the conditional block that checks for the absence of errors in the data. This enhancement prevents unnecessary calculations when errors are present in the data, ensuring more efficient processing.

refactor: Optimize Payment Status-Based Calculation

This commit optimizes the code by refining the calculation process based on the payment status.
Specifically, the calculation is now performed only when the payment status is 'OK', eliminating unnecessary calculations when the payment status is not equal to 'OK'. This enhancement improves the efficiency of the code without altering its functionality.

fix: Ensure Correct Package Setting in OrderAdmin

This fix addresses an issue in the `save_form` method within the `OrderAdmin` class. The method is responsible for creating instances of the `Order` model. The issue was related to the execution of the `set_package` method. Specifically, when the 'completed' field in the form's cleaned data had a True value and the user possessed the necessary permission ('payment.change_completed_order'), the `set_package` method was not being executed as intended.

The commit rectifies this by adjusting the condition to accurately check for the 'completed' field in the cleaned data using `form.cleaned_data.get('completed')`. This ensures that the `set_package` method is executed only when appropriate, based on the accurate value of the 'completed' field and the user's permissions.

feat: Add Package Validation in Order Serializer

This enhancement introduces a `validate_package` method within the `OrderCreateOrUpdateSerializer` to ensure accurate validation of the 'package' field. This method confirms that the provided package value is valid and fulfills specific criteria.

The commit improves data integrity by verifying that the selected package is not marked for deletion(is_delete=False) and has received confirmation(confirmation=True). In case of an invalid package value, a `serializers.ValidationError` is raised to prevent erroneous data entry.

This strategic addition strengthens the reliability of the `OrderCreateOrUpdateSerializer` by enforcing stringent validation rules for the 'package' field.

refactor: Add 'pk' field in 'fields' of OrderReadSerializer's Meta Class