fix: Ensure Token Activation After Successful Ad Creation
This commit addresses an issue in the `CreateAdAPI` view that occurred when an error was raised during the `ser.save()` operation. Previously, if an error occurred during the `ser.save()` operation, the `token_activated` field was set to `False`, even though an ad might not have been created.

To rectify this issue, the commit revises the order of operations. Now, after the `ser.save()` operation successfully creates an ad, the `token_activated` field is set to `False`. This change ensures that the `token_activated` field remains `True` until a successful ad creation has been confirmed.